Added new tabs to simulation: RelativeRisks, Scenarios, Diseases.

+++ Dynamo-HIA_UI/src/nl/rivm/emi/dynamo/estimation/InitialPopulationFactory.java
@@ -566,7 +566,7 @@ public class InitialPopulationFactory {
 										+ individualNumber + "_ref");
 						individualNumber++;
 						// long seed2 = (long) rand2.random();
-						long seed2 =(((long)rand2.random())<<32)+(long)rand2.random();
+						long seed2 = (long) ( ((long)rand2.random() ) << 32)+ (long)rand2.random();
 						currentIndividual.setRandomNumberGeneratorSeed(seed2);
 
 						/*
